FEAT: Se corregio los parametros erroneos en el registro de ventas que surgieron el code review
